Add bid confirmation to gallery page, bump version, and remove unnecessary !important
- Add confirm() dialog to gallery card quick-bid handler (was missing, only
  single-item and modal handlers had it)
- Bump version to 0.9.6 to bust browser cache for updated aih-frontend.js
- Fix PHP 8.1 float-to-int deprecation in bid timestamp milliseconds
- Remove ~40 unnecessary !important declarations in elegant-theme.css,
  using increased selector specificity (.aih-page prefix) instead
- Move inline styles from single-item.php to proper CSS class

+++ includes/class-aih-bid.php
@@ -33,7 +33,7 @@ class AIH_Bid {
         
         $art_table = AIH_Database::get_table('art_pieces');
         $now = current_time('mysql');
-        $ms = substr(sprintf('%03d', (microtime(true) * 1000) % 1000), 0, 3);
+        $ms = substr(sprintf('%03d', (int)((microtime(true) * 1000) % 1000)), 0, 3);
         $bid_time = $now . '.' . $ms;
 
         // Start transaction to prevent race conditions
